Update PlutusIR to match the API of PlutusCore
- add PlutusIR.Core module,
- move AST declarations to PlutusIR.Core.Type,
- move traversals to PlutusIR.Core.Plated,
- move instance declarations to Plutus.Core.Instance,
- move pretty printing instances to PlutusIR.Core.Instance.Pretty,
- add PlutusIR.Mark,
- add PlutusIR.Subst,
- restrict exports from PlutusIR.Transform.Inline,
- use setOf from Data.Set.Lens,
- implement missing bits of PlutusIR API.
