+++ nix/pkgs/haskell/materialized-unix/.plan.nix/plutus-core.nix
@@ -243,14 +243,21 @@
           "PlutusCore/Parser"
           "PlutusCore/Error"
           "PlutusIR"
+          "PlutusIR/Core"
+          "PlutusIR/Core/Instance"
+          "PlutusIR/Core/Instance/Pretty"
+          "PlutusIR/Core/Plated"
+          "PlutusIR/Core/Type"
           "PlutusIR/Compiler"
           "PlutusIR/Compiler/Names"
           "PlutusIR/Compiler/Definitions"
           "PlutusIR/Error"
           "PlutusIR/Generators/AST"
           "PlutusIR/Parser"
+          "PlutusIR/Mark"
           "PlutusIR/MkPir"
           "PlutusIR/Purity"
+          "PlutusIR/Subst"
           "PlutusIR/Transform/DeadCode"
           "PlutusIR/Transform/Substitute"
           "PlutusIR/Transform/ThunkRecursions"
