+++ plutus-core/plutus-core.cabal
@@ -99,14 +99,21 @@ library
         PlutusCore.Error
 
         PlutusIR
+        PlutusIR.Core
+        PlutusIR.Core.Instance
+        PlutusIR.Core.Instance.Pretty
+        PlutusIR.Core.Plated
+        PlutusIR.Core.Type
         PlutusIR.Compiler
         PlutusIR.Compiler.Names
         PlutusIR.Compiler.Definitions
         PlutusIR.Error
         PlutusIR.Generators.AST
         PlutusIR.Parser
+        PlutusIR.Mark
         PlutusIR.MkPir
         PlutusIR.Purity
+        PlutusIR.Subst
         PlutusIR.Transform.DeadCode
         PlutusIR.Transform.Substitute
         PlutusIR.Transform.ThunkRecursions
