+++ plutus-core/plutus-core/src/PlutusCore/Subst.hs
@@ -29,6 +29,7 @@ import           PlutusCore.Name
 import           Control.Lens
 import           Data.Functor.Foldable (cata)
 import           Data.Set              as Set
+import           Data.Set.Lens         (setOf)
 
 purely :: ((a -> Identity b) -> c -> Identity d) -> (a -> b) -> c -> d
 purely = coerce
@@ -183,11 +184,6 @@ ftvTy = cata f
     f (TyLamF _ bnd _ ty)    = delete bnd ty
     f (TyAppF _ ty1 ty2)     = ty1 `union` ty2
     f TyBuiltinF{}           = Set.empty
-
--- All variables
-
-setOf :: Getting (Set a) s a -> s -> Set a
-setOf g = foldMapOf g singleton
 
 -- | Get all the term variables in a term.
 vTerm :: Ord name => Term tyname name uni fun ann -> Set name
