+++ plutus-core/plutus-ir/src/PlutusIR/Core/Instance.hs
@@ -1,0 +1,3 @@
+module PlutusIR.Core.Instance () where
+
+import           PlutusIR.Core.Instance.Pretty ()
