+++ plutus-core/untyped-plutus-core/src/UntypedPlutusCore/Subst.hs
@@ -22,6 +22,7 @@ import           PlutusCore.Name
 
 import           Control.Lens
 import           Data.Set               as Set
+import           Data.Set.Lens          (setOf)
 
 purely :: ((a -> Identity b) -> c -> Identity d) -> (a -> b) -> c -> d
 purely = coerce
@@ -112,11 +113,6 @@ programMapNames
     -> Program name uni fun ann
     -> Program name' uni fun ann
 programMapNames f (Program a v term) = Program a v (termMapNames f term)
-
--- All variables
-
-setOf :: Getting (Set a) s a -> s -> Set a
-setOf g = foldMapOf g singleton
 
 -- | Get all the term variables in a term.
 vTerm :: Ord name => Term name uni fun ann -> Set name
